+++ apps/www/app/layout.tsx
@@ -5,7 +5,7 @@ import { SpeedInsights } from "@vercel/speed-insights/next";
 import { RootProvider } from "fumadocs-ui/provider/next";
 import type { Metadata } from "next";
 import { Inter, JetBrains_Mono } from "next/font/google";
-import { type ReactNode, useId } from "react";
+import type { ReactNode } from "react";
 import { Toaster } from "~/components/ui/toaster";
 
 const inter = Inter({
@@ -48,7 +48,7 @@ export default function Layout({ children }: { children: ReactNode }) {
 						enableSystem: true,
 					}}
 				>
-					<Banner variant="rainbow" id={useId()}>
+					<Banner variant="rainbow" id="arktype-feature-banner">
 						🎉 We are now featured on&nbsp;
 						<a
 							href="https://arktype.io/docs/ecosystem#arkenv"