+++ packages/arkenv/src/errors.ts
@@ -1,5 +1,5 @@
 import type { ArkErrors } from "arktype";
-import { styleText } from "./style";
+import { styleText } from "./lib/style-text";
 import { indent } from "./utils";
 
 /**